requirements.txt
streamlit
PyPDF2
pytesseract
Pillow
beautifulsoup4
requests
rdflib
python-docx
matplotlib 
plotly 
altair
tensorflow
